chore: Reorganize the token provider
Apply a newer provider structure to keep modules smaller and better
structured.
