+++ benches/fernet_token.rs
@@ -7,8 +7,8 @@ use std::io::Write;
 use tempfile::tempdir;
 
 use openstack_keystone::config::Config;
-use openstack_keystone::token::fernet::FernetTokenProvider;
-use openstack_keystone::token::fernet::bench_get_fernet_timestamp;
+use openstack_keystone::token::backend::fernet::FernetTokenProvider;
+use openstack_keystone::token::backend::fernet::bench_get_fernet_timestamp;
 //use openstack_keystone::token::types::TokenBackend;
 
 fn decode(backend: &FernetTokenProvider, token: &str) {
