+++ src/token/backend/fernet/restricted.rs
@@ -11,68 +11,18 @@
 // limitations under the License.
 //
 // SPDX-License-Identifier: Apache-2.0
+//! Restricted token Fernet implementation.
 
-use chrono::{DateTime, Utc};
-use derive_builder::Builder;
 use rmp::{decode::read_pfix, encode::write_pfix};
-use serde::Serialize;
 use std::io::Write;
 
-use crate::identity::types::UserResponse;
 use crate::token::{
+    backend::fernet::{FernetTokenProvider, MsgPackToken, utils},
     error::TokenProviderError,
-    fernet::{FernetTokenProvider, MsgPackToken},
-    fernet_utils,
-    types::Token,
+    types::RestrictedPayload,
 };
 
-#[derive(Builder, Clone, Debug, Default, PartialEq, Serialize)]
-#[builder(setter(into))]
-pub struct UnscopedPayload {
-    pub user_id: String,
-    #[builder(default, setter(name = _methods))]
-    pub methods: Vec<String>,
-    #[builder(default, setter(name = _audit_ids))]
-    pub audit_ids: Vec<String>,
-    pub expires_at: DateTime<Utc>,
-
-    #[builder(default)]
-    pub issued_at: DateTime<Utc>,
-    #[builder(default)]
-    pub user: Option<UserResponse>,
-}
-
-impl UnscopedPayloadBuilder {
-    pub fn methods<I, V>(&mut self, iter: I) -> &mut Self
-    where
-        I: Iterator<Item = V>,
-        V: Into<String>,
-    {
-        self.methods
-            .get_or_insert_with(Vec::new)
-            .extend(iter.map(Into::into));
-        self
-    }
-
-    pub fn audit_ids<I, V>(&mut self, iter: I) -> &mut Self
-    where
-        I: Iterator<Item = V>,
-        V: Into<String>,
-    {
-        self.audit_ids
-            .get_or_insert_with(Vec::new)
-            .extend(iter.map(Into::into));
-        self
-    }
-}
-
-impl From<UnscopedPayload> for Token {
-    fn from(value: UnscopedPayload) -> Self {
-        Self::Unscoped(value)
-    }
-}
-
-impl MsgPackToken for UnscopedPayload {
+impl MsgPackToken for RestrictedPayload {
     type Token = Self;
 
     fn assemble<W: Write>(
@@ -80,14 +30,18 @@ impl MsgPackToken for UnscopedPayload {
         wd: &mut W,
         fernet_provider: &FernetTokenProvider,
     ) -> Result<(), TokenProviderError> {
-        fernet_utils::write_uuid(wd, &self.user_id)?;
+        utils::write_uuid(wd, &self.user_id)?;
         write_pfix(
             wd,
             fernet_provider.encode_auth_methods(self.methods.clone())?,
         )
         .map_err(|x| TokenProviderError::RmpEncode(x.to_string()))?;
-        fernet_utils::write_time(wd, self.expires_at)?;
-        fernet_utils::write_audit_ids(wd, self.audit_ids.clone())?;
+        utils::write_uuid(wd, &self.token_restriction_id)?;
+        utils::write_time(wd, self.expires_at)?;
+        utils::write_uuid(wd, &self.project_id)?;
+        utils::write_bool(wd, self.allow_renew)?;
+        utils::write_bool(wd, self.allow_rescope)?;
+        utils::write_audit_ids(wd, self.audit_ids.clone())?;
 
         Ok(())
     }
@@ -96,19 +50,28 @@ impl MsgPackToken for UnscopedPayload {
         rd: &mut &[u8],
         fernet_provider: &FernetTokenProvider,
     ) -> Result<Self::Token, TokenProviderError> {
-        // Order of writing is important
-        let user_id = fernet_utils::read_uuid(rd)?;
+        // Order of reading is important
+        let user_id = utils::read_uuid(rd)?;
         let methods: Vec<String> = fernet_provider
             .decode_auth_methods(read_pfix(rd)?)?
             .into_iter()
             .collect();
-        let expires_at = fernet_utils::read_time(rd)?;
-        let audit_ids: Vec<String> = fernet_utils::read_audit_ids(rd)?.into_iter().collect();
-        Ok(Self::Token {
+        let token_restriction_id = utils::read_uuid(rd)?;
+        let expires_at = utils::read_time(rd)?;
+        let project_id = utils::read_uuid(rd)?;
+        let allow_renew = utils::read_bool(rd)?;
+        let allow_rescope = utils::read_bool(rd)?;
+        let audit_ids: Vec<String> = utils::read_audit_ids(rd)?.into_iter().collect();
+        Ok(Self {
             user_id,
             methods,
             expires_at,
             audit_ids,
+            token_restriction_id,
+            project_id,
+            allow_renew,
+            allow_rescope,
+
             ..Default::default()
         })
     }
@@ -119,16 +82,20 @@ mod tests {
     use chrono::{Local, SubsecRound};
     use uuid::Uuid;
 
-    use super::super::tests::setup_config;
     use super::*;
+    use crate::token::tests::setup_config;
 
     #[test]
     fn test_roundtrip() {
-        let token = UnscopedPayload {
+        let token = RestrictedPayload {
             user_id: Uuid::new_v4().simple().to_string(),
-            methods: vec!["password".into()],
+            methods: vec!["openid".into()],
             audit_ids: vec!["Zm9vCg".into()],
             expires_at: Local::now().trunc_subsecs(0).into(),
+            token_restriction_id: "trid".into(),
+            project_id: "pid".into(),
+            allow_renew: true,
+            allow_rescope: true,
             ..Default::default()
         };
 
@@ -137,7 +104,8 @@ mod tests {
         let mut buf = vec![];
         token.assemble(&mut buf, &provider).unwrap();
         let encoded_buf = buf.clone();
-        let decoded = UnscopedPayload::disassemble(&mut encoded_buf.as_slice(), &provider).unwrap();
+        let decoded =
+            RestrictedPayload::disassemble(&mut encoded_buf.as_slice(), &provider).unwrap();
         assert_eq!(token, decoded);
     }
 }
