+++ src/token/types/application_credential.rs
@@ -1,0 +1,74 @@
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+//
+// SPDX-License-Identifier: Apache-2.0
+
+use chrono::{DateTime, Utc};
+use derive_builder::Builder;
+use serde::Serialize;
+
+use crate::assignment::types::Role;
+use crate::identity::types::UserResponse;
+use crate::resource::types::Project;
+use crate::token::types::Token;
+
+#[derive(Builder, Clone, Debug, Default, PartialEq, Serialize)]
+#[builder(setter(into))]
+pub struct ApplicationCredentialPayload {
+    pub user_id: String,
+    #[builder(default, setter(name = _methods))]
+    pub methods: Vec<String>,
+    #[builder(default, setter(name = _audit_ids))]
+    pub audit_ids: Vec<String>,
+    pub expires_at: DateTime<Utc>,
+    pub project_id: String,
+    pub application_credential_id: String,
+
+    #[builder(default)]
+    pub issued_at: DateTime<Utc>,
+    #[builder(default)]
+    pub user: Option<UserResponse>,
+    #[builder(default)]
+    pub roles: Vec<Role>,
+    #[builder(default)]
+    pub project: Option<Project>,
+}
+
+impl ApplicationCredentialPayloadBuilder {
+    pub fn methods<I, V>(&mut self, iter: I) -> &mut Self
+    where
+        I: Iterator<Item = V>,
+        V: Into<String>,
+    {
+        self.methods
+            .get_or_insert_with(Vec::new)
+            .extend(iter.map(Into::into));
+        self
+    }
+
+    pub fn audit_ids<I, V>(&mut self, iter: I) -> &mut Self
+    where
+        I: Iterator<Item = V>,
+        V: Into<String>,
+    {
+        self.audit_ids
+            .get_or_insert_with(Vec::new)
+            .extend(iter.map(Into::into));
+        self
+    }
+}
+
+impl From<ApplicationCredentialPayload> for Token {
+    fn from(value: ApplicationCredentialPayload) -> Self {
+        Self::ApplicationCredential(value)
+    }
+}
