+++ src/token/types/federation_project_scoped.rs
@@ -22,34 +22,25 @@ use crate::assignment::types::Role;
 use crate::identity::types::UserResponse;
 use crate::resource::types::Project;
 use crate::token::{
+    backend::fernet::{FernetTokenProvider, MsgPackToken, utils},
     error::TokenProviderError,
-    fernet::{FernetTokenProvider, MsgPackToken},
-    fernet_utils,
     types::Token,
 };
 
-/// Restricted token payload
+/// Federated project scope token payload
 #[derive(Builder, Clone, Debug, Default, PartialEq, Serialize)]
 #[builder(setter(into))]
-pub struct RestrictedPayload {
-    /// User ID.
+pub struct FederationProjectScopePayload {
     pub user_id: String,
-    /// Authentication methods used to obtain the token.
     #[builder(default, setter(name = _methods))]
     pub methods: Vec<String>,
-    /// Token audit IDs.
     #[builder(default, setter(name = _audit_ids))]
     pub audit_ids: Vec<String>,
-    /// Token expiration datetime in UTC.
     pub expires_at: DateTime<Utc>,
-    /// ID of the token restrictions.
-    pub token_restriction_id: String,
-    /// Project ID scope for the token.
     pub project_id: String,
-    /// Whether the token can be renewed.
-    pub allow_renew: bool,
-    /// Whether the token can be rescoped.
-    pub allow_rescope: bool,
+    pub idp_id: String,
+    pub protocol_id: String,
+    pub group_ids: Vec<String>,
 
     #[builder(default)]
     pub issued_at: DateTime<Utc>,
@@ -61,7 +52,7 @@ pub struct RestrictedPayload {
     pub project: Option<Project>,
 }
 
-impl RestrictedPayloadBuilder {
+impl FederationProjectScopePayloadBuilder {
     pub fn methods<I, V>(&mut self, iter: I) -> &mut Self
     where
         I: Iterator<Item = V>,
@@ -85,13 +76,13 @@ impl RestrictedPayloadBuilder {
     }
 }
 
-impl From<RestrictedPayload> for Token {
-    fn from(value: RestrictedPayload) -> Self {
-        Self::Restricted(value)
+impl From<FederationProjectScopePayload> for Token {
+    fn from(value: FederationProjectScopePayload) -> Self {
+        Self::FederationProjectScope(value)
     }
 }
 
-impl MsgPackToken for RestrictedPayload {
+impl MsgPackToken for FederationProjectScopePayload {
     type Token = Self;
 
     fn assemble<W: Write>(
@@ -99,18 +90,18 @@ impl MsgPackToken for RestrictedPayload {
         wd: &mut W,
         fernet_provider: &FernetTokenProvider,
     ) -> Result<(), TokenProviderError> {
-        fernet_utils::write_uuid(wd, &self.user_id)?;
+        utils::write_uuid(wd, &self.user_id)?;
         write_pfix(
             wd,
             fernet_provider.encode_auth_methods(self.methods.clone())?,
         )
         .map_err(|x| TokenProviderError::RmpEncode(x.to_string()))?;
-        fernet_utils::write_uuid(wd, &self.token_restriction_id)?;
-        fernet_utils::write_time(wd, self.expires_at)?;
-        fernet_utils::write_uuid(wd, &self.project_id)?;
-        fernet_utils::write_bool(wd, self.allow_renew)?;
-        fernet_utils::write_bool(wd, self.allow_rescope)?;
-        fernet_utils::write_audit_ids(wd, self.audit_ids.clone())?;
+        utils::write_uuid(wd, &self.project_id)?;
+        utils::write_list_of_uuids(wd, self.group_ids.iter())?;
+        utils::write_uuid(wd, &self.idp_id)?;
+        utils::write_str(wd, &self.protocol_id)?;
+        utils::write_time(wd, self.expires_at)?;
+        utils::write_audit_ids(wd, self.audit_ids.clone())?;
 
         Ok(())
     }
@@ -120,27 +111,26 @@ impl MsgPackToken for RestrictedPayload {
         fernet_provider: &FernetTokenProvider,
     ) -> Result<Self::Token, TokenProviderError> {
         // Order of reading is important
-        let user_id = fernet_utils::read_uuid(rd)?;
+        let user_id = utils::read_uuid(rd)?;
         let methods: Vec<String> = fernet_provider
             .decode_auth_methods(read_pfix(rd)?)?
             .into_iter()
             .collect();
-        let token_restriction_id = fernet_utils::read_uuid(rd)?;
-        let expires_at = fernet_utils::read_time(rd)?;
-        let project_id = fernet_utils::read_uuid(rd)?;
-        let allow_renew = fernet_utils::read_bool(rd)?;
-        let allow_rescope = fernet_utils::read_bool(rd)?;
-        let audit_ids: Vec<String> = fernet_utils::read_audit_ids(rd)?.into_iter().collect();
+        let project_id = utils::read_uuid(rd)?;
+        let group_ids = utils::read_list_of_uuids(rd)?;
+        let idp_id = utils::read_uuid(rd)?;
+        let protocol_id = utils::read_str(rd)?;
+        let expires_at = utils::read_time(rd)?;
+        let audit_ids: Vec<String> = utils::read_audit_ids(rd)?.into_iter().collect();
         Ok(Self {
             user_id,
             methods,
             expires_at,
             audit_ids,
-            token_restriction_id,
             project_id,
-            allow_renew,
-            allow_rescope,
-
+            group_ids: group_ids.into_iter().collect(),
+            idp_id,
+            protocol_id,
             ..Default::default()
         })
     }
@@ -151,20 +141,20 @@ mod tests {
     use chrono::{Local, SubsecRound};
     use uuid::Uuid;
 
-    use super::super::tests::setup_config;
     use super::*;
+    use crate::token::tests::setup_config;
 
     #[test]
     fn test_roundtrip() {
-        let token = RestrictedPayload {
+        let token = FederationProjectScopePayload {
             user_id: Uuid::new_v4().simple().to_string(),
             methods: vec!["openid".into()],
             audit_ids: vec!["Zm9vCg".into()],
             expires_at: Local::now().trunc_subsecs(0).into(),
-            token_restriction_id: "trid".into(),
             project_id: "pid".into(),
-            allow_renew: true,
-            allow_rescope: true,
+            group_ids: vec!["g1".into()],
+            idp_id: "idp_id".into(),
+            protocol_id: "proto".into(),
             ..Default::default()
         };
 
@@ -174,7 +164,8 @@ mod tests {
         token.assemble(&mut buf, &provider).unwrap();
         let encoded_buf = buf.clone();
         let decoded =
-            RestrictedPayload::disassemble(&mut encoded_buf.as_slice(), &provider).unwrap();
+            FederationProjectScopePayload::disassemble(&mut encoded_buf.as_slice(), &provider)
+                .unwrap();
         assert_eq!(token, decoded);
     }
 }
