+++ src/token/types/federation_unscoped.rs
@@ -18,27 +18,23 @@ use rmp::{decode::read_pfix, encode::write_pfix};
 use serde::Serialize;
 use std::io::Write;
 
-use crate::assignment::types::Role;
 use crate::identity::types::UserResponse;
-use crate::resource::types::Project;
 use crate::token::{
+    backend::fernet::{FernetTokenProvider, MsgPackToken, utils},
     error::TokenProviderError,
-    fernet::{FernetTokenProvider, MsgPackToken},
-    fernet_utils,
     types::Token,
 };
 
-/// Federated project scope token payload
+/// Federated unscoped token payload
 #[derive(Builder, Clone, Debug, Default, PartialEq, Serialize)]
 #[builder(setter(into))]
-pub struct FederationProjectScopePayload {
+pub struct FederationUnscopedPayload {
     pub user_id: String,
     #[builder(default, setter(name = _methods))]
     pub methods: Vec<String>,
     #[builder(default, setter(name = _audit_ids))]
     pub audit_ids: Vec<String>,
     pub expires_at: DateTime<Utc>,
-    pub project_id: String,
     pub idp_id: String,
     pub protocol_id: String,
     pub group_ids: Vec<String>,
@@ -47,13 +43,9 @@ pub struct FederationProjectScopePayload {
     pub issued_at: DateTime<Utc>,
     #[builder(default)]
     pub user: Option<UserResponse>,
-    #[builder(default)]
-    pub roles: Option<Vec<Role>>,
-    #[builder(default)]
-    pub project: Option<Project>,
 }
 
-impl FederationProjectScopePayloadBuilder {
+impl FederationUnscopedPayloadBuilder {
     pub fn methods<I, V>(&mut self, iter: I) -> &mut Self
     where
         I: Iterator<Item = V>,
@@ -77,13 +69,13 @@ impl FederationProjectScopePayloadBuilder {
     }
 }
 
-impl From<FederationProjectScopePayload> for Token {
-    fn from(value: FederationProjectScopePayload) -> Self {
-        Self::FederationProjectScope(value)
+impl From<FederationUnscopedPayload> for Token {
+    fn from(value: FederationUnscopedPayload) -> Self {
+        Self::FederationUnscoped(value)
     }
 }
 
-impl MsgPackToken for FederationProjectScopePayload {
+impl MsgPackToken for FederationUnscopedPayload {
     type Token = Self;
 
     fn assemble<W: Write>(
@@ -91,18 +83,17 @@ impl MsgPackToken for FederationProjectScopePayload {
         wd: &mut W,
         fernet_provider: &FernetTokenProvider,
     ) -> Result<(), TokenProviderError> {
-        fernet_utils::write_uuid(wd, &self.user_id)?;
+        utils::write_uuid(wd, &self.user_id)?;
         write_pfix(
             wd,
             fernet_provider.encode_auth_methods(self.methods.clone())?,
         )
         .map_err(|x| TokenProviderError::RmpEncode(x.to_string()))?;
-        fernet_utils::write_uuid(wd, &self.project_id)?;
-        fernet_utils::write_list_of_uuids(wd, self.group_ids.iter())?;
-        fernet_utils::write_uuid(wd, &self.idp_id)?;
-        fernet_utils::write_str(wd, &self.protocol_id)?;
-        fernet_utils::write_time(wd, self.expires_at)?;
-        fernet_utils::write_audit_ids(wd, self.audit_ids.clone())?;
+        utils::write_list_of_uuids(wd, self.group_ids.iter())?;
+        utils::write_uuid(wd, &self.idp_id)?;
+        utils::write_str(wd, &self.protocol_id)?;
+        utils::write_time(wd, self.expires_at)?;
+        utils::write_audit_ids(wd, self.audit_ids.clone())?;
 
         Ok(())
     }
@@ -112,23 +103,21 @@ impl MsgPackToken for FederationProjectScopePayload {
         fernet_provider: &FernetTokenProvider,
     ) -> Result<Self::Token, TokenProviderError> {
         // Order of reading is important
-        let user_id = fernet_utils::read_uuid(rd)?;
+        let user_id = utils::read_uuid(rd)?;
         let methods: Vec<String> = fernet_provider
             .decode_auth_methods(read_pfix(rd)?)?
             .into_iter()
             .collect();
-        let project_id = fernet_utils::read_uuid(rd)?;
-        let group_ids = fernet_utils::read_list_of_uuids(rd)?;
-        let idp_id = fernet_utils::read_uuid(rd)?;
-        let protocol_id = fernet_utils::read_str(rd)?;
-        let expires_at = fernet_utils::read_time(rd)?;
-        let audit_ids: Vec<String> = fernet_utils::read_audit_ids(rd)?.into_iter().collect();
+        let group_ids = utils::read_list_of_uuids(rd)?;
+        let idp_id = utils::read_uuid(rd)?;
+        let protocol_id = utils::read_str(rd)?;
+        let expires_at = utils::read_time(rd)?;
+        let audit_ids: Vec<String> = utils::read_audit_ids(rd)?.into_iter().collect();
         Ok(Self {
             user_id,
             methods,
             expires_at,
             audit_ids,
-            project_id,
             group_ids: group_ids.into_iter().collect(),
             idp_id,
             protocol_id,
@@ -142,17 +131,16 @@ mod tests {
     use chrono::{Local, SubsecRound};
     use uuid::Uuid;
 
-    use super::super::tests::setup_config;
     use super::*;
+    use crate::token::tests::setup_config;
 
     #[test]
     fn test_roundtrip() {
-        let token = FederationProjectScopePayload {
+        let token = FederationUnscopedPayload {
             user_id: Uuid::new_v4().simple().to_string(),
             methods: vec!["openid".into()],
             audit_ids: vec!["Zm9vCg".into()],
             expires_at: Local::now().trunc_subsecs(0).into(),
-            project_id: "pid".into(),
             group_ids: vec!["g1".into()],
             idp_id: "idp_id".into(),
             protocol_id: "proto".into(),
@@ -165,8 +153,7 @@ mod tests {
         token.assemble(&mut buf, &provider).unwrap();
         let encoded_buf = buf.clone();
         let decoded =
-            FederationProjectScopePayload::disassemble(&mut encoded_buf.as_slice(), &provider)
-                .unwrap();
+            FederationUnscopedPayload::disassemble(&mut encoded_buf.as_slice(), &provider).unwrap();
         assert_eq!(token, decoded);
     }
 }
