+++ src/token/types/project_scoped.rs
@@ -22,9 +22,8 @@ use crate::assignment::types::Role;
 use crate::identity::types::UserResponse;
 use crate::resource::types::Project;
 use crate::token::{
+    backend::fernet::{FernetTokenProvider, MsgPackToken, utils},
     error::TokenProviderError,
-    fernet::{FernetTokenProvider, MsgPackToken},
-    fernet_utils,
     types::Token,
 };
 
@@ -87,15 +86,15 @@ impl MsgPackToken for ProjectScopePayload {
         wd: &mut W,
         fernet_provider: &FernetTokenProvider,
     ) -> Result<(), TokenProviderError> {
-        fernet_utils::write_uuid(wd, &self.user_id)?;
+        utils::write_uuid(wd, &self.user_id)?;
         write_pfix(
             wd,
             fernet_provider.encode_auth_methods(self.methods.clone())?,
         )
         .map_err(|x| TokenProviderError::RmpEncode(x.to_string()))?;
-        fernet_utils::write_uuid(wd, &self.project_id)?;
-        fernet_utils::write_time(wd, self.expires_at)?;
-        fernet_utils::write_audit_ids(wd, self.audit_ids.clone())?;
+        utils::write_uuid(wd, &self.project_id)?;
+        utils::write_time(wd, self.expires_at)?;
+        utils::write_audit_ids(wd, self.audit_ids.clone())?;
 
         Ok(())
     }
@@ -105,14 +104,14 @@ impl MsgPackToken for ProjectScopePayload {
         fernet_provider: &FernetTokenProvider,
     ) -> Result<Self::Token, TokenProviderError> {
         // Order of reading is important
-        let user_id = fernet_utils::read_uuid(rd)?;
+        let user_id = utils::read_uuid(rd)?;
         let methods: Vec<String> = fernet_provider
             .decode_auth_methods(read_pfix(rd)?)?
             .into_iter()
             .collect();
-        let project_id = fernet_utils::read_uuid(rd)?;
-        let expires_at = fernet_utils::read_time(rd)?;
-        let audit_ids: Vec<String> = fernet_utils::read_audit_ids(rd)?.into_iter().collect();
+        let project_id = utils::read_uuid(rd)?;
+        let expires_at = utils::read_time(rd)?;
+        let audit_ids: Vec<String> = utils::read_audit_ids(rd)?.into_iter().collect();
         Ok(Self {
             user_id,
             methods,
@@ -129,8 +128,8 @@ mod tests {
     use chrono::{Local, SubsecRound};
     use uuid::Uuid;
 
-    use super::super::tests::setup_config;
     use super::*;
+    use crate::token::tests::setup_config;
 
     #[test]
     fn test_roundtrip() {
