+++ src/token/types/unscoped.rs
@@ -20,25 +20,20 @@ use std::io::Write;
 
 use crate::identity::types::UserResponse;
 use crate::token::{
+    backend::fernet::{FernetTokenProvider, MsgPackToken, utils},
     error::TokenProviderError,
-    fernet::{FernetTokenProvider, MsgPackToken},
-    fernet_utils,
     types::Token,
 };
 
-/// Federated unscoped token payload
 #[derive(Builder, Clone, Debug, Default, PartialEq, Serialize)]
 #[builder(setter(into))]
-pub struct FederationUnscopedPayload {
+pub struct UnscopedPayload {
     pub user_id: String,
     #[builder(default, setter(name = _methods))]
     pub methods: Vec<String>,
     #[builder(default, setter(name = _audit_ids))]
     pub audit_ids: Vec<String>,
     pub expires_at: DateTime<Utc>,
-    pub idp_id: String,
-    pub protocol_id: String,
-    pub group_ids: Vec<String>,
 
     #[builder(default)]
     pub issued_at: DateTime<Utc>,
@@ -46,7 +41,7 @@ pub struct FederationUnscopedPayload {
     pub user: Option<UserResponse>,
 }
 
-impl FederationUnscopedPayloadBuilder {
+impl UnscopedPayloadBuilder {
     pub fn methods<I, V>(&mut self, iter: I) -> &mut Self
     where
         I: Iterator<Item = V>,
@@ -70,13 +65,13 @@ impl FederationUnscopedPayloadBuilder {
     }
 }
 
-impl From<FederationUnscopedPayload> for Token {
-    fn from(value: FederationUnscopedPayload) -> Self {
-        Self::FederationUnscoped(value)
+impl From<UnscopedPayload> for Token {
+    fn from(value: UnscopedPayload) -> Self {
+        Self::Unscoped(value)
     }
 }
 
-impl MsgPackToken for FederationUnscopedPayload {
+impl MsgPackToken for UnscopedPayload {
     type Token = Self;
 
     fn assemble<W: Write>(
@@ -84,17 +79,14 @@ impl MsgPackToken for FederationUnscopedPayload {
         wd: &mut W,
         fernet_provider: &FernetTokenProvider,
     ) -> Result<(), TokenProviderError> {
-        fernet_utils::write_uuid(wd, &self.user_id)?;
+        utils::write_uuid(wd, &self.user_id)?;
         write_pfix(
             wd,
             fernet_provider.encode_auth_methods(self.methods.clone())?,
         )
         .map_err(|x| TokenProviderError::RmpEncode(x.to_string()))?;
-        fernet_utils::write_list_of_uuids(wd, self.group_ids.iter())?;
-        fernet_utils::write_uuid(wd, &self.idp_id)?;
-        fernet_utils::write_str(wd, &self.protocol_id)?;
-        fernet_utils::write_time(wd, self.expires_at)?;
-        fernet_utils::write_audit_ids(wd, self.audit_ids.clone())?;
+        utils::write_time(wd, self.expires_at)?;
+        utils::write_audit_ids(wd, self.audit_ids.clone())?;
 
         Ok(())
     }
@@ -103,25 +95,19 @@ impl MsgPackToken for FederationUnscopedPayload {
         rd: &mut &[u8],
         fernet_provider: &FernetTokenProvider,
     ) -> Result<Self::Token, TokenProviderError> {
-        // Order of reading is important
-        let user_id = fernet_utils::read_uuid(rd)?;
+        // Order of writing is important
+        let user_id = utils::read_uuid(rd)?;
         let methods: Vec<String> = fernet_provider
             .decode_auth_methods(read_pfix(rd)?)?
             .into_iter()
             .collect();
-        let group_ids = fernet_utils::read_list_of_uuids(rd)?;
-        let idp_id = fernet_utils::read_uuid(rd)?;
-        let protocol_id = fernet_utils::read_str(rd)?;
-        let expires_at = fernet_utils::read_time(rd)?;
-        let audit_ids: Vec<String> = fernet_utils::read_audit_ids(rd)?.into_iter().collect();
-        Ok(Self {
+        let expires_at = utils::read_time(rd)?;
+        let audit_ids: Vec<String> = utils::read_audit_ids(rd)?.into_iter().collect();
+        Ok(Self::Token {
             user_id,
             methods,
             expires_at,
             audit_ids,
-            group_ids: group_ids.into_iter().collect(),
-            idp_id,
-            protocol_id,
             ..Default::default()
         })
     }
@@ -132,19 +118,16 @@ mod tests {
     use chrono::{Local, SubsecRound};
     use uuid::Uuid;
 
-    use super::super::tests::setup_config;
     use super::*;
+    use crate::token::tests::setup_config;
 
     #[test]
     fn test_roundtrip() {
-        let token = FederationUnscopedPayload {
+        let token = UnscopedPayload {
             user_id: Uuid::new_v4().simple().to_string(),
-            methods: vec!["openid".into()],
+            methods: vec!["password".into()],
             audit_ids: vec!["Zm9vCg".into()],
             expires_at: Local::now().trunc_subsecs(0).into(),
-            group_ids: vec!["g1".into()],
-            idp_id: "idp_id".into(),
-            protocol_id: "proto".into(),
             ..Default::default()
         };
 
@@ -153,8 +136,7 @@ mod tests {
         let mut buf = vec![];
         token.assemble(&mut buf, &provider).unwrap();
         let encoded_buf = buf.clone();
-        let decoded =
-            FederationUnscopedPayload::disassemble(&mut encoded_buf.as_slice(), &provider).unwrap();
+        let decoded = UnscopedPayload::disassemble(&mut encoded_buf.as_slice(), &provider).unwrap();
         assert_eq!(token, decoded);
     }
 }
